calendar, in case int is 16-bit

+++ src/c99/calendar.c
@@ -94,7 +94,7 @@ double novas_jd_from_date(enum novas_calendar_type calendar, int year, int month
     return NAN;
   }
 
-  jd = day - 32123 + 1461 * (year + 4800L + m14 / 12) / 4 + 367 * (month - 2 - m14 / 12 * 12) / 12;
+  jd = day - 32123 + 1461LL * (year + 4800L + m14 / 12) / 4 + 367 * (month - 2 - m14 / 12 * 12) / 12;
   fjd = (hour - 12.0) / DAY_HOURS;
 
   if(calendar == NOVAS_ASTRONOMICAL_CALENDAR)
@@ -171,18 +171,18 @@ int novas_jd_to_date(double tjd, enum novas_calendar_type calendar, int *restric
   if(h < 0.0)
     h += 24.0;
 
-  k = jd + 68569;
-  n = 4 * k / 146097;
+  k = jd + 68569L;
+  n = 4 * k / 146097L;
 
   if(calendar == NOVAS_ASTRONOMICAL_CALENDAR)
     calendar = (tjd >= NOVAS_JD_START_GREGORIAN) ? NOVAS_GREGORIAN_CALENDAR : NOVAS_ROMAN_CALENDAR;
 
   if(calendar == NOVAS_GREGORIAN_CALENDAR)
-    k -= (146097 * n + 3) / 4;
+    k -= (146097L * n + 3) / 4;
   else
-    k -= (146100 * n + 3) / 4;
+    k -= (146100L * n + 3) / 4;
 
-  m = 4000 * (k + 1) / 1461001;
+  m = 4000 * (k + 1) / 1461001L;
 
   k += 31 - 1461 * m / 4;
 
@@ -190,10 +190,10 @@ int novas_jd_to_date(double tjd, enum novas_calendar_type calendar, int *restric
     k += 38;
 
   mo = (int) (80 * k / 2447);
-  d = (int) (k - 2447 * (long) mo / 80);
+  d = (int) (k - 2447L * mo / 80);
   k = mo / 11;
 
-  mo = (int) ((long) mo + 2 - 12 * k);
+  mo = (int) (mo + 2 - 12 * k);
   y = (int) (100 * (n - 49) + m + k);
 
   if(year)
